Booking itinerary,activity adding to it a field needed which is time_to_attend and handling cases ,Delete activity and itinerary added check if some on booked it cannot delete

+++ API/src/services/itineraryService.ts
@@ -150,6 +150,13 @@ export default class ItineraryService {
     );
     if (!tourGuide) throw new HttpError("Tour guide not found", 404);
 
+    const tickets = await this.ticketModel.find({ booking_id: itinerary_id });
+
+    if (tickets.length > 0)
+      throw new BadRequestError(
+        "Itinerary booked by some users so cannot delete"
+      );
+
     const deletedItinerary = await this.itineraryModel.findByIdAndDelete(
       itinerary_id
     );
